gestures

+++ Components/FavouriteComponent.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import { View, Text } from 'react-native';
+import { View, Text, Alert } from 'react-native';
 import {ListItem} from 'react-native-elements';
 import { baseUrl } from '../shared/baseUrl';
 import Loading from './LoadingComponent';
@@ -7,6 +7,8 @@ import { FlatList } from 'react-native-gesture-handler';
 import { connect } from 'react-redux';
 import Swipeout from 'react-native-swipeout';
 import {deleteFavorite} from '../redux/ActionCreator';
+import * as Animatable from 'react-native-animatable';
+
 
 const mapStateToProps = state => {
     return {
@@ -28,22 +30,41 @@ class Favorites extends Component {
         const renderMenuItem = ({item, index}) => {
             const rightButton = [
                 {
-                    text : 'Delete',
-                    type : 'delete',
-                    onPress : () => this.props.deleteFavorite(item.id)
+                    text: 'Delete', 
+                    type: 'delete',
+                    onPress: () => {
+                        Alert.alert(
+                            'Delete Favorite?',
+                            'Are you sure you wish to delete the favorite dish ' + item.name + '?',
+                            [
+                                { 
+                                    text: 'Cancel', 
+                                    onPress: () => console.log(item.name + 'Not Deleted'),
+                                    style: ' cancel'
+                                },
+                                {
+                                    text: 'OK',
+                                    onPress: () => this.props.deleteFavorite(item.id)
+                                }
+                            ],
+                            { cancelable: false }
+                        );
+                        
+                    }
                 }
             ]
             return (
-                <Swipeout right = {rightButton} autoClose>
-                    <ListItem
-                        key={index}
-                        title={item.name}
-                        subtitle={item.description}
-                        hideChevron={true}
-                        onPress={() => navigate('DishDetail', { dishId: item.id })}
-                        leftAvatar={{ source: {uri: baseUrl + item.image}}}
-                    />
-                </Swipeout>
+                
+                    <Swipeout right = {rightButton} autoClose>
+                        <ListItem
+                            key={index}
+                            title={item.name}
+                            subtitle={item.description}
+                            hideChevron={true}
+                            onPress={() => navigate('DishDetail', { dishId: item.id })}
+                            leftAvatar={{ source: {uri: baseUrl + item.image}}}
+                        />
+                    </Swipeout>
             );
         };
 
@@ -62,11 +83,13 @@ class Favorites extends Component {
         else {
             if(this.props.favorites.length){
                 return (
-                    <FlatList 
-                        data={this.props.dishes.dishes.filter(dish => this.props.favorites.some(el => el === dish.id))}
-                        renderItem={renderMenuItem}
-                        keyExtractor={item => item.id.toString()}
-                    />
+                    <Animatable.View animation="fadeInRightBig" duration={2000}>                
+                        <FlatList 
+                            data={this.props.dishes.dishes.filter(dish => this.props.favorites.some(el => el === dish.id))}
+                            renderItem={renderMenuItem}
+                            keyExtractor={item => item.id.toString()}
+                        />
+                    </Animatable.View>
                 );
             }
             else{
@@ -78,8 +101,6 @@ class Favorites extends Component {
                     </View>
                 );
             }
-            
-            
         }
     }
 }
